Bag fix ver 3.1

+++ lab03/src/Factorial.java
@@ -1,14 +1,14 @@
 public class Factorial {
 
-    public  int recursionExecute(int n){
+    public long recursionExecute(long n){
         if (n < 2) {
                 return 1;
             }
         return recursionExecute(n - 1) * n;
     }
 
-    public  int cycleExecute(int n){
-        int result = 1;
+    public long cycleExecute(int n){
+        long result = 1;
 
         for (int i = 1; i <= n; i ++){
